requirements.txt
boto3==1.26.137
botocore==1.29.137
pathlib==1.0.1
python-dotenv==1.0.0
flask==2.0.1
gunicorn==20.1.0 